secure-kyc-chain/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pillow==11.0.0
numpy==1.26.4
scikit-learn==1.5.1
shap==0.45.1
opencv-python==4.10.0.84
deepface==0.0.79
paddlepaddle==2.6.1
paddleocr==2.7.3
insightface==0.7.3
onnxruntime==1.16.3
dlib==19.24.2
httpx==0.27.2
alembic==1.13.2
pytest==8.3.3
pytest-asyncio==0.24.0

